0.9.7.1 - Significant update to low level panel interaction
0.9.7.1
Over the past 6 years the low level code has had fixes on fixes on fixes and has become unmanageable.  This is the first release of low level code that centralises the sequencing and control of the panels interaction.
As such, although I believe that I have maintained these fixes, I can't guarantee that 100% so be prepared to raise Issues on Github