0.9.7.8 - English Translation and some Bugs Squashed
So more bugs squashed, the arming/disarming commands should be much more responsive now. Selecting "Panel" as the language should be the same as selecting English when in Standard Mode or the EPROM is not Downloaded.
Most of this release though is to do with maturing the English Translation. I know what you're thinking, it's in English anyway.  I have significantly improved en.json in the translations directory.  This means that you should all be able to take that and create other language translations e.g fr.json for French, ge.json for German. Just leave the text to the left of the colon and translate the right text string to the language.