0.9.7.9 - English Translation File, Logic Updates and Bug Fixes
English Translation File, Logic Updates and Bug Fixes.
The English language file is 99% complete, and I can't think of what the 1% could be but I've probably missed something. All possible settings for all states and attributes (when they are strings and not numbers) are included. So for example, you can see what all the possible lasteventname and lasteventaction values could be.
Remaining Bug: Motion/Camera PIRs on PowerMaster panels can miss triggers.  There is an improvement on previous releases but it still misses a few.
I have removed the FR and NL language translations temporarily until I can do language files for them. However, there will still be a lot in English as it was never complete anyway.