0.9.7.10 - Update to fix Powermax+ Panel
I changed the startup sequence but unfortunately it didn't work with Powermax+ panels. So I've had to put some of the startup sequence back for all panels (when the integration starts it doesn't know the panel type so I can't make any assumptions).  The impact for all panels is that it may take about 10 seconds longer to reach Powerlink Emulation Mode.
I have tested it with my own Powermax Pro Part and my test PowerMaster30 and it seems to be stable.